fix to analysis: verify if outputGrid is valid, changing band value from int to size_t

+++ src/terrama2/services/analysis/core/grid/forecast/Operator.hpp
@@ -63,7 +63,7 @@ namespace terrama2
               \return A double value with the result.
             */
             double operatorImpl(terrama2::services::analysis::core::StatisticOperation statisticOperation,
-                                const std::string& dataSeriesName, const std::string& dateFilterEnd, const int band);
+                                const std::string& dataSeriesName, const std::string& dateFilterEnd, const size_t band);
 
             /*!
               \brief Calculates the sum of historic grid data.
@@ -75,7 +75,7 @@ namespace terrama2
 
               \return A double value with the result.
             */
-            double sum(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double sum(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the mean of historic grid data.
@@ -87,7 +87,7 @@ namespace terrama2
 
               \return A double value with the result.
             */
-            double mean(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double mean(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the min of historic grid data.
@@ -99,7 +99,7 @@ namespace terrama2
 
               \return A double value with the result.
             */
-            double min(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double min(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the max of historic grid data.
@@ -114,7 +114,7 @@ namespace terrama2
 
               \return A double value with the result.
             */
-            double max(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double max(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the median of historic grid data.
@@ -126,7 +126,7 @@ namespace terrama2
 
               \return A double value with the result.
             */
-            double median(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double median(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the standard deviation of historic grid data.
@@ -137,7 +137,7 @@ namespace terrama2
               \param dateFilter Time filter for the data.
               \return A double value with the result.
             */
-            double standardDeviation(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double standardDeviation(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
             /*!
               \brief Calculates the variance of historic grid data.
@@ -148,7 +148,7 @@ namespace terrama2
               \param dateFilter Time filter for the data.
               \return A double value with the result.
             */
-            double variance(const std::string& dataSeriesName, const std::string& dateFilter, const int band = 0);
+            double variance(const std::string& dataSeriesName, const std::string& dateFilter, const size_t band = 0);
 
           } // end namespace forecast
         }   // end namespace grid
