+++ src/terrama2/services/analysis/core/grid/zonal/history/prec/Operator.hpp
@@ -126,7 +126,7 @@ namespace terrama2
                                     const std::string& dataSeriesName,
                                     const std::string& dateDiscardBefore,
                                     const std::string& dateDiscardAfter,
-                                    const int band,
+                                    const size_t band,
                                     terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 double count(const std::string& dataSeriesName, const std::string& dateDiscardBefore, terrama2::services::analysis::core::Buffer buffer = Buffer());
@@ -140,7 +140,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double sum(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double sum(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the mean of zonal grid data.
@@ -153,7 +153,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double mean(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double mean(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the min of zonal grid data.
@@ -165,7 +165,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double min(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double min(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the max of zonal grid data.
@@ -180,7 +180,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double max(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double max(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the median of zonal grid data.
@@ -192,7 +192,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double median(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double median(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the standard deviation of zonal grid data.
@@ -203,7 +203,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double standardDeviation(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double standardDeviation(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the variance of zonal grid data.
@@ -214,7 +214,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double variance(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double variance(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 void appendValues(te::rst::Raster* raster, te::gm::Polygon* polygon, std::vector<double>& values);
               } /* prec */
