+++ src/terrama2/services/analysis/core/grid/zonal/history/ratio/Operator.hpp
@@ -71,7 +71,7 @@ namespace terrama2
                                     const std::string& dataSeriesName,
                                     const std::string& dateDiscardBefore,
                                     const std::string& dateDiscardAfter,
-                                    const int band,
+                                    const size_t band,
                                     terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
@@ -94,7 +94,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double sum(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double sum(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the mean of zonal grid data.
@@ -106,7 +106,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double mean(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double mean(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the min of zonal grid data.
@@ -118,7 +118,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double min(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double min(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the max of zonal grid data.
@@ -133,7 +133,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double max(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double max(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the median of zonal grid data.
@@ -145,7 +145,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double median(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double median(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the standard deviation of zonal grid data.
@@ -156,7 +156,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double standardDeviation(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double standardDeviation(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
 
                 /*!
                   \brief Calculates the variance of zonal grid data.
@@ -167,7 +167,7 @@ namespace terrama2
 
                   \return A double value with the result.
                 */
-                double variance(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const int band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
+                double variance(const std::string& dataSeriesName, const std::string& dateDiscardBefore, const size_t band = 0, terrama2::services::analysis::core::Buffer buffer = Buffer());
               } /* ratio */
             }
           } /* zonal */
